+++ CHANGES.md
@@ -5,6 +5,7 @@ Release Notes.
 0.4.0
 ------------------
 #### Features
+* Add support ignore suffix.
 
 #### Plugins
 * Support setting a discard type of reporter.
@@ -21,6 +22,7 @@ Release Notes.
 * Fix ParseVendorModule error for special case in vendor/modules.txt.
 * Fix enhance method error when unknown parameter type.
 * Fix wrong tracing context when trace have been sampled.
+* Fix enhance param error when there are multiple params.
 
 #### Issues and PR
 - All issues are [here](https://github.com/apache/skywalking/milestone/197?closed=1)